+++ tests/_test.py
@@ -2,15 +2,18 @@
 tests._test
 ===========
 """
+import pytest
+
 from object_colors import Color
 
-from . import ATTRS, GREEN, INSTANCES, RED, RESET, TEST_STR
+from . import ATTRS, CODES, INSTANCES, RESET, TEST_STR, TEST_TUPLE
 
 
-def test_color_string():
+@pytest.mark.parametrize("fore,expected", [("red", 1), ("green", 2)])
+def test_color_string(fore, expected):
     """Test a simple string."""
-    color = Color(fore="green")
-    assert color.get(TEST_STR) == f"{GREEN}{TEST_STR}{RESET}"
+    color = Color(fore=fore)
+    assert color.get(TEST_STR) == f"{CODES[expected]}{TEST_STR}{RESET}"
 
 
 def test__getattr__(color):
@@ -71,7 +74,10 @@ def test_color_print(capsys, populated_colors):
     """
     populated_colors.red.print(TEST_STR)
     captured = capsys.readouterr()
-    assert captured.out == f"{RED}{TEST_STR}{RESET}\n"
+    assert captured.out == f"{CODES[1]}{TEST_STR}{RESET}\n"
+    populated_colors.red.print(*TEST_TUPLE)
+    captured = capsys.readouterr()
+    assert captured.out == f"{CODES[1]}{' '.join(TEST_TUPLE)}{RESET}\n"
 
 
 def test_repr(color, capsys) -> None:
@@ -87,3 +93,27 @@ def test_repr(color, capsys) -> None:
         "Color(effect=0, fore=7, back=None, bold=Color(effect=1, fore=7, "
         "back=None))"
     )
+
+
+def test_set_subclass(capsys):
+    """Test that a color subclass can be set within the main
+    instance.
+    """
+    color = Color()
+    color.set(the_name_is_up_to_the_user={"fore": "red"})
+    color.the_name_is_up_to_the_user.print(TEST_STR)
+    captured = capsys.readouterr()
+    assert captured.out == f"{CODES[1]}{TEST_STR}{RESET}\n"
+
+
+def test_set_regular_attr(capsys):
+    """Test that an attr can be set with the ``set`` method too
+    instance.
+    """
+    color = Color()
+    assert color.fore == 7
+    color.set(fore="red")
+    assert color.fore == 1
+    color.print(TEST_STR)
+    captured = capsys.readouterr()
+    assert captured.out == f"{CODES[1]}{TEST_STR}{RESET}\n"